backend/requirements.txt
fastapi
uvicorn
websockets
psycopg2-binary
python-dotenv
sqlalchemy 
faker